+++ src/main/java/com/gistpetition/api/DataLoader.java
@@ -1,7 +1,11 @@
 package com.gistpetition.api;
 
 import com.gistpetition.api.petition.application.PetitionCommandService;
-import com.gistpetition.api.petition.domain.*;
+import com.gistpetition.api.petition.domain.Category;
+import com.gistpetition.api.petition.domain.Petition;
+import com.gistpetition.api.petition.domain.repository.AgreementRepository;
+import com.gistpetition.api.petition.domain.repository.AnswerRepository;
+import com.gistpetition.api.petition.domain.repository.PetitionRepository;
 import com.gistpetition.api.petition.dto.request.AgreementRequest;
 import com.gistpetition.api.petition.dto.request.AnswerRequest;
 import com.gistpetition.api.petition.dto.request.PetitionRequest;
