+++ src/main/java/com/gistpetition/api/petition/application/PetitionQueryService.java
@@ -3,7 +3,12 @@ package com.gistpetition.api.petition.application;
 import com.gistpetition.api.exception.petition.NoSuchPetitionException;
 import com.gistpetition.api.exception.petition.NotReleasedPetitionException;
 import com.gistpetition.api.exception.user.NoSuchUserException;
-import com.gistpetition.api.petition.domain.*;
+import com.gistpetition.api.petition.domain.Agreement;
+import com.gistpetition.api.petition.domain.Answer;
+import com.gistpetition.api.petition.domain.Category;
+import com.gistpetition.api.petition.domain.Petition;
+import com.gistpetition.api.petition.domain.repository.AgreementRepository;
+import com.gistpetition.api.petition.domain.repository.PetitionRepository;
 import com.gistpetition.api.petition.dto.response.AgreementResponse;
 import com.gistpetition.api.petition.dto.response.PetitionPreviewResponse;
 import com.gistpetition.api.petition.dto.response.PetitionResponse;
