+++ src/test/java/com/gistpetition/api/petition/application/PetitionServiceTest.java
@@ -4,7 +4,11 @@ import com.gistpetition.api.ServiceTest;
 import com.gistpetition.api.exception.petition.DuplicatedAgreementException;
 import com.gistpetition.api.exception.petition.NoSuchPetitionException;
 import com.gistpetition.api.petition.PetitionBuilder;
-import com.gistpetition.api.petition.domain.*;
+import com.gistpetition.api.petition.domain.Agreement;
+import com.gistpetition.api.petition.domain.Category;
+import com.gistpetition.api.petition.domain.Petition;
+import com.gistpetition.api.petition.domain.repository.AgreementRepository;
+import com.gistpetition.api.petition.domain.repository.PetitionRepository;
 import com.gistpetition.api.petition.dto.request.AgreementRequest;
 import com.gistpetition.api.petition.dto.request.AnswerRequest;
 import com.gistpetition.api.petition.dto.request.PetitionRequest;
